backend/requirements.txt
fastapi
uvicorn
python-multipart
easyocr
scikit-learn
joblib
numpy
pandas
requests
deep-translator
pillow
google-generativeai
python-dotenv
